refactor code with class Impl
