+++ modules/wechat_qrcode/src/detector/align.hpp
