+++ modules/wechat_qrcode/src/scale/super_scale.hpp
